Update seed snapshots
