+++ seed/swift-sdk/single-url-environment-default/with-custom-config/Sources/Requests/Requests.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+/// Container for all inline request types used throughout the SDK.
+///
+/// This enum serves as a namespace to organize request types that are defined inline within endpoint specifications.
+public enum Requests {}
